TS-35531 Adapted system test to work in CI
