+++ system-tests/gradle-cucumber/gradle-project/settings.gradle.kts
@@ -1,9 +1,12 @@
 pluginManagement {
     repositories {
         mavenLocal()
-        mavenCentral()
         gradlePluginPortal()
     }
+}
+
+plugins {
+	id("org.gradle.toolchains.foojay-resolver-convention") version "0.5.0"
 }
 
 rootProject.name = "gradle-cucumber"
